requirements.txt
numpy
tensorflow-gpu==1.8.0
foolbox==1.1.0
adversarial-vision-challenge

